Using qt 6.4.2

+++ libqt6c/src/libqbuffer.hpp
@@ -34,7 +34,7 @@ int QBuffer_Metacall(QBuffer* self, int param1, int param2, void** param3);
 libqt_string QBuffer_Buffer(QBuffer* self);
 libqt_string QBuffer_Buffer2(const QBuffer* self);
 void QBuffer_SetData(QBuffer* self, const libqt_string data);
-void QBuffer_SetData2(QBuffer* self, const char* data, ptrdiff_t lenVal);
+void QBuffer_SetData2(QBuffer* self, const char* data, int lenVal);
 libqt_string QBuffer_Data(const QBuffer* self);
 bool QBuffer_Open(QBuffer* self, int openMode);
 void QBuffer_Close(QBuffer* self);
